+++ Source/ACE.Server/WorldObjects/Creature_Death.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+
 using ACE.Database;
+using ACE.Database.Models.Shard;
 using ACE.Database.Models.World;
 using ACE.Entity;
 using ACE.Entity.Enum;
@@ -161,7 +163,7 @@ namespace ACE.Server.WorldObjects
         {
             if (NoCorpse)
             {
-                var loot = GenerateTreasure(null);
+                var loot = GenerateTreasure(killer, null);
 
                 foreach(var item in loot)
                 {
@@ -226,7 +228,7 @@ namespace ACE.Server.WorldObjects
             // set 'killed by' for looting rights
             if (killer != null)
             {
-                corpse.LongDesc = $"Killed by {killer.Name.TrimStart('+')}."; // VTANK requires + to be stripped for regex matching.
+                corpse.LongDesc = $"Killed by {killer.Name.TrimStart('+')}.";  // vtank requires + to be stripped for regex matching.
                 if (killer is CombatPet)
                     corpse.KillerId = killer.PetOwner.Value;
                 else
@@ -258,7 +260,7 @@ namespace ACE.Server.WorldObjects
             else
             {
                 corpse.IsMonster = true;
-                GenerateTreasure(corpse);
+                GenerateTreasure(killer, corpse);
                 if (killer is Player && (Level >= 100 || Level >= killer.Level + 5))
                 {
                     CanGenerateRare = true;
@@ -293,7 +295,7 @@ namespace ACE.Server.WorldObjects
         /// <summary>
         /// Transfers generated treasure from creature to corpse
         /// </summary>
-        private List<WorldObject> GenerateTreasure(Corpse corpse)
+        private List<WorldObject> GenerateTreasure(WorldObject killer, Corpse corpse)
         {
             var droppedItems = new List<WorldObject>();
 
@@ -307,6 +309,8 @@ namespace ACE.Server.WorldObjects
                         corpse.TryAddToInventory(wo);
                     else
                         droppedItems.Add(wo);
+
+                    corpse.DoCantripLogging(killer, wo);
                 }
             }