+++ Source/ACE.Server/WorldObjects/Creature_Tick.cs
@@ -1,4 +1,8 @@
+using System.Collections.Generic;
 using System.Linq;
+
+using ACE.Entity.Enum;
+using ACE.Server.Network.GameMessages.Messages;
 
 namespace ACE.Server.WorldObjects
 {
@@ -9,8 +13,10 @@ namespace ACE.Server.WorldObjects
         /// </summary>
         public override void Heartbeat(double currentUnixTime)
         {
+            var expireItems = new List<WorldObject>();
+
             // added where clause
-            foreach (var wo in EquippedObjects.Values.Where(i => i.EnchantmentManager.HasEnchantments))
+            foreach (var wo in EquippedObjects.Values.Where(i => i.EnchantmentManager.HasEnchantments || i.RemainingLifespan.HasValue))
             {
                 // FIXME: wo.NextHeartbeatTime is double.MaxValue here
                 //if (wo.NextHeartbeatTime <= currentUnixTime)
@@ -18,8 +24,15 @@ namespace ACE.Server.WorldObjects
 
                 // just go by parent heartbeats, only for enchantments?
                 // TODO: handle players dropping / picking up items
-                wo.EnchantmentManager.HeartBeat(HeartbeatInterval);
+                wo.EnchantmentManager.HeartBeat(CachedHeartbeatInterval);
 
+                if (wo.RemainingLifespan != null)
+                {
+                    wo.RemainingLifespan -= (int)CachedHeartbeatInterval;
+
+                    if (wo.RemainingLifespan <= 0)
+                        expireItems.Add(wo);
+                }
             }
 
             VitalHeartBeat();
@@ -27,6 +40,15 @@ namespace ACE.Server.WorldObjects
             EmoteManager.HeartBeat();
 
             DamageHistory.TryPrune();
+
+            // delete items when RemainingLifespan <= 0
+            foreach (var expireItem in expireItems)
+            {
+                expireItem.DeleteObject(this);
+
+                if (this is Player player)
+                    player.Session.Network.EnqueueSend(new GameMessageSystemChat($"Its lifespan finished, your {expireItem.Name} crumbles to dust.", ChatMessageType.Broadcast));
+            }
 
             base.Heartbeat(currentUnixTime);
         }