+++ Source/ACE.Server/WorldObjects/Player_Combat.cs
@@ -618,17 +618,27 @@ namespace ACE.Server.WorldObjects
         /// <summary>
         /// Returns TRUE if this player is PK and died to another player
         /// </summary>
-        public bool IsPKDeath()
+        public bool IsPKDeath(WorldObject topDamager)
         {
-            return PlayerKillerStatus.HasFlag(PlayerKillerStatus.PK) && new ObjectGuid(KillerId ?? 0).IsPlayer();
+            return IsPKDeath(topDamager?.Guid.Full);
+        }
+
+        public bool IsPKDeath(uint? killerGuid)
+        {
+            return PlayerKillerStatus.HasFlag(PlayerKillerStatus.PK) && new ObjectGuid(killerGuid ?? 0).IsPlayer();
         }
 
         /// <summary>
         /// Returns TRUE if this player is PKLite and died to another player
         /// </summary>
-        public bool IsPKLiteDeath()
+        public bool IsPKLiteDeath(WorldObject topDamager)
         {
-            return PlayerKillerStatus.HasFlag(PlayerKillerStatus.PKLite) && new ObjectGuid(KillerId ?? 0).IsPlayer();
+            return IsPKLiteDeath(topDamager?.Guid.Full);
+        }
+
+        public bool IsPKLiteDeath(uint? killerGuid)
+        {
+            return PlayerKillerStatus.HasFlag(PlayerKillerStatus.PKLite) && new ObjectGuid(killerGuid ?? 0).IsPlayer();
         }
 
         /// <summary>