+++ Source/ACE.Server/WorldObjects/Player_Skills.cs
@@ -6,6 +6,7 @@ using ACE.DatLoader;
 using ACE.Entity.Enum;
 using ACE.Entity.Enum.Properties;
 using ACE.Server.Entity.Actions;
+using ACE.Server.Managers;
 using ACE.Server.Network.GameMessages.Messages;
 using ACE.Server.WorldObjects.Entity;
 
@@ -551,7 +552,17 @@ namespace ACE.Server.WorldObjects
 
         public override bool GetHeritageBonus(WorldObject weapon)
         {
-            return GetHeritageBonus(GetWeaponType(weapon));
+            if (weapon == null || !weapon.IsMasterable)
+                return false;
+
+            if (PropertyManager.GetBool("universal_masteries").Item)
+            {
+                // https://asheron.fandom.com/wiki/Spring_2014_Update
+                // end of retail - universal masteries
+                return true;
+            }
+            else
+                return GetHeritageBonus(GetWeaponType(weapon));
         }
 
         public bool GetHeritageBonus(WeaponType weaponType)