+++ Source/ACE.Server/WorldObjects/WorldObject_Magic.cs
@@ -5,16 +5,17 @@ using System.Numerics;
 using System.Text;
 
 using ACE.Database;
+using ACE.Database.Models.Shard;
 using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Entity.Enum.Properties;
 using ACE.Server.Entity;
+using ACE.Server.Factories;
 using ACE.Server.Network.GameEvent.Events;
 using ACE.Server.Network.GameMessages.Messages;
 using ACE.Server.Network.Structure;
 using ACE.Server.Managers;
 using ACE.Server.Entity.Actions;
-using ACE.Server.Factories;
 using ACE.Server.Physics;
 
 namespace ACE.Server.WorldObjects
@@ -1880,5 +1881,15 @@ namespace ACE.Server.WorldObjects
                     return ResistanceType.Undef;
             }
         }
+
+        /// <summary>
+        /// Returns the epic cantrips from this item's spellbook
+        /// </summary>
+        public List<BiotaPropertiesSpellBook> EpicCantrips => Biota.BiotaPropertiesSpellBook.Where(i => LootTables.EpicCantrips.Contains(i.Spell)).ToList();
+
+        /// <summary>
+        /// Returns the legendary cantrips from this item's spellbook
+        /// </summary>
+        public List<BiotaPropertiesSpellBook> LegendaryCantrips => Biota.BiotaPropertiesSpellBook.Where(i => LootTables.LegendaryCantrips.Contains(i.Spell)).ToList();
     }
 }